backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
pgvector==0.2.4
openai==1.10.0
langchain==0.1.4
langchain-community==0.0.16
langchain-openai==0.0.5
pinecone-client==3.0.0
tiktoken==0.5.2
pydantic==2.6.0
pydantic-settings==2.1.0
requests==2.31.0
aiofiles==23.2.1
